+++ RgbLeds.h
@@ -1,18 +1,18 @@
 #ifndef RgbLeds_h
 #define RgbLeds_h
 
-#include "Arduino.h"
+#include <Arduino.h>
 
 typedef std::function<void(uint8_t, int)> AnalogWriteFunction;
 typedef std::function<void(uint8_t, uint8_t)> PinModeFunction;
 
 class RgbLeds {
 public:
-  const uint redPin;
-  const uint greenPin;
-  const uint bluePin;
+  const uint8_t redPin;
+  const uint8_t greenPin;
+  const uint8_t bluePin;
 
-  RgbLeds(uint _redPin, uint _greenPin, uint _bluePin);
+  RgbLeds(uint8_t _redPin, uint8_t _greenPin, uint8_t _bluePin);
 
   void init(PinModeFunction pinMode, AnalogWriteFunction analogWrite);
   String toString();
